+++ src/main/java/org/sopt/config/SecurityConfig.java
@@ -1,9 +1,10 @@
 package org.sopt.config;
 
 import lombok.RequiredArgsConstructor;
+import org.sopt.security.JwtAuthFilter;
+import org.sopt.security.SecurityErrorHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
